Removes environment manager
