+++ index.html
@@ -5,7 +5,6 @@
         <title>Zombie</title>
         <script src="./dist/vendor/phaser.js"></script>
         <script src="./dist/js/main.bundle.js"></script>
-        <link rel="stylesheet" href="./dist/css/main.css">
         <style>
             body {
                 margin: 0;
